Create SimpleSnake
SNAKE is a math game where you roll 3 dice, multiply the numbers shown, and then choose whether to stop and keep your score or to roll again to potentially add more. If at some point you roll all the same numbers, you loose all the points for that round. There are 5 rounds, same as there are 5 letters in "SNAKE," hence the name.